Added new functionality: dynamic log directory
Log dir can now be specified in experiment configuration file (under section Log). If the field log > dir is left empty, then logs will be saved in a study-specific directory within the directory where the code resides. In case a directory is specified, then a study-specific directory is created (if necessary) within this directory and the log files are written here.